ISIS-2441: finally move ObjectAdapter to JDO (no longer used in core)
