--- core/metamodel/src/main/java/org/apache/isis/core/metamodel/adapter/package-info.java
@@ -1,32 +1,0 @@
-/*
- *  Licensed to the Apache Software Foundation (ASF) under one
- *  or more contributor license agreements.  See the NOTICE file
- *  distributed with this work for additional information
- *  regarding copyright ownership.  The ASF licenses this file
- *  to you under the Apache License, Version 2.0 (the
- *  "License"); you may not use this file except in compliance
- *  with the License.  You may obtain a copy of the License at
- *
- *        http://www.apache.org/licenses/LICENSE-2.0
- *
- *  Unless required by applicable law or agreed to in writing,
- *  software distributed under the License is distributed on an
- *  "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- *  KIND, either express or implied.  See the License for the
- *  specific language governing permissions and limitations
- *  under the License.
- */
-
-/**
- * Defines {@link org.apache.isis.core.metamodel.adapter.ObjectAdapter} interface,
- * a wrapper around each domain object pojo.
- *
- * <p>
- * {@link org.apache.isis.core.metamodel.adapter.ObjectAdapter} is analogous
- * to {@link java.lang.Object java.lang.Object}; from it the {@link org.apache.isis.core.metamodel.spec.ObjectSpecification}
- * can be obtained (analogous to {@link java.lang.Class java.lang.Class}.
- *
- * <p>
- * The other significant type defined by
- */
-package org.apache.isis.core.metamodel.adapter;
