+++ persistence/jdo/datanucleus-5/src/main/java/org/apache/isis/persistence/jdo/datanucleus5/objectadapter/ObjectAdapterProvider.java
@@ -20,8 +20,6 @@ package org.apache.isis.persistence.jdo.datanucleus5.objectadapter;
 
 import javax.annotation.Nullable;
 
-import org.apache.isis.core.metamodel.adapter.ObjectAdapter;
-
 /**
  * 
  * @since 2.0
